feat: system messages style: row 0 gap default refactor: close button moved
