+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "upvite",
-  "version": "0.2.19",
+  "version": "0.2.20",
   "private": true,
   "scripts": {
     "dev": "vite dev",
